openafs-kmod: Only build kernel module
I've removed everything related to the openafs project except for
building the kernel module.

Added to this is a new executable, find-installed-kversion.sh, which
uses the installed kernel headers to define the kversion, so the build
process no longer uses the output of `uname -r` or a --define
parameter to the rpmbuild command.  Instead, it parses
LINUX_VERSION_CODE and RHEL_RELEASE defined in <linux/version.h> and
builds the kversion from this.

Also, we're using the Red Hat kmodtool script instead of the OpenAFS
one.  It looks like this might make it easier to build dependencies
against kernel symbols instead of particular kernel releases.