openafs-kmod.spec: add 'kernel' as a BuildRequires
The Red Hat rpmbuild automatic dependency generator uses the kernel's
symbols to add kernel symbol dependencies to the package.  For
example, if the kernel module uses vprintk(), it might use add this as
a dependency to the kmod package:

kernel(vprintk) = 0xbf8ba54a

To get access to the kernel symbols, the rpmbuild automatic dependency
generator needs the 'kernel' package installed.